backend/requirements.txt
# =========================================================
# Dependencias del Backend - Sistema de Inventario Kardex
# Instalar con: pip install -r requirements.txt
# =========================================================

# Framework web
fastapi==0.115.0
uvicorn[standard]==0.30.0

# ORM y base de datos
sqlalchemy==2.0.35
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Validación y configuración
pydantic-settings==2.5.0
python-dotenv==1.0.1

# Autenticación JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# Migraciones de base de datos
alembic==1.13.0

# Testing
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0

# Utilidades
python-multipart==0.0.9
